// frontend/src/api/api.ts
export interface CurrentUserResponseDto {
    /**
     * 
     * @type {CurrentUserStatus}
     * @memberof CurrentUserResponseDto
     */
    'status'?: CurrentUserStatus;
    /**
     * Login username.
     * @type {string}
     * @memberof CurrentUserResponseDto
     */
    'userName'?: string | null;
    /**
     * Email address.
     * @type {string}
     * @memberof CurrentUserResponseDto
     */
    'email'?: string | null;
    /**
     * List of roles associated to this user.
     * @type {Array<string>}
     * @memberof CurrentUserResponseDto
     */
    'roles'?: Array<string> | null;
}


/**
 * M:ExampleWebApp.Backend.WebApi.AuthController.CurrentUser response api specific status.  0 = OK  2 = InvalidArgument  3 = InvalidAuthentication  100 = AccessTokenNotFound
 * @export
 * @enum {string}
 */

export const CurrentUserStatus = {
    /**
    * 
    */
    OK: 0,
    /**
    * 
    */
    InvalidArgument: 2,
    /**
    * 
    */
    InvalidAuthentication: 3,
    /**
    * 
    */
    AccessTokenNotFound: 100
} as const;

export type CurrentUserStatus = typeof CurrentUserStatus[keyof typeof CurrentUserStatus];
// ...
export interface LoginResponseDto {
    /**
     * 
     * @type {LoginStatus}
     * @memberof LoginResponseDto
     */
    'status'?: LoginStatus;
    /**
     * Username.
     * @type {string}
     * @memberof LoginResponseDto
     */
    'userName'?: string | null;
    /**
     * Email.
     * @type {string}
     * @memberof LoginResponseDto
     */
    'email'?: string | null;
    /**
     * User roles.
     * @type {Array<string>}
     * @memberof LoginResponseDto
     */
    'roles'?: Array<string> | null;
}


/**
 * M:ExampleWebApp.Backend.WebApi.AuthController.Login(ExampleWebApp.Backend.WebApi.LoginRequestDto) response api specific status.  0 = OK  2 = UsernameOrEmailRequired  3 = InvalidAuthentication  8 = InvalidHttpContext
 * @export
 * @enum {string}
 */

export const LoginStatus = {
    /**
    * 
    */
    OK: 0,
    /**
    * 
    */
    UsernameOrEmailRequired: 2,
    /**
    * 
    */
    InvalidAuthentication: 3,
    /**
    * 
    */
    InvalidHttpContext: 8
} as const;

export type LoginStatus = typeof LoginStatus[keyof typeof LoginStatus];
// ...
export interface UserListItemResponseDto {
    /**
     * User name.
     * @type {string}
     * @memberof UserListItemResponseDto
     */
    'userName'?: string | null;
    /**
     * User email.
     * @type {string}
     * @memberof UserListItemResponseDto
     */
    'email'?: string | null;
    /**
     * User roles.
     * @type {Array<string>}
     * @memberof UserListItemResponseDto
     */
    'roles'?: Array<string> | null;
    /**
     * Access failed count.
     * @type {number}
     * @memberof UserListItemResponseDto
     */
    'accessFailedCount'?: number;
    /**
     * Email is confirmed.
     * @type {boolean}
     * @memberof UserListItemResponseDto
     */
    'emailConfirmed'?: boolean;
    /**
     * Lockout end (UTC).
     * @type {string}
     * @memberof UserListItemResponseDto
     */
    'lockoutEnd'?: string | null;
    /**
     * User phone number.
     * @type {string}
     * @memberof UserListItemResponseDto
     */
    'phoneNumber'?: string | null;
    /**
     * User phone number confirmed.
     * @type {boolean}
     * @memberof UserListItemResponseDto
     */
    'phoneNumberConfirmed'?: boolean;
    /**
     * Two factor enabled.
     * @type {boolean}
     * @memberof UserListItemResponseDto
     */
    'twoFactorEnabled'?: boolean;
}